Fixes stack level to deep error
The synchronisation dialog seems very sensible since it does not render
in production but raises the stack level error instead.

After playing around in development the error was reproduceble. Removing
the presenter around the 'details' partial solve the problem. It seems
as if the presenter is not appropriate in order to execute plain ruby
code, i.e., without view helpers.

Fixes #1176