Considers admin setting for link copied issues
An admin can configure the linking behavior of copied issues. Up to now
it is ignored to be considered in Issue#copy_from. Instead there is
a separate method ProjectPatch#establish_relation where the admin
setting is considered. But this is too late because the relation 'copied_*'
will be set in an issue callback after copying. That is, it already exist when
ariving ProjectPatch#establish_relation.

Fixes #1728