Sorts out links to issues on other projects
When an issue was synchronized to another project before it gets synchronized
again then it might have links to the copy in the other project.

Usually this 'copied_to' link will be synchronized too. This behavior
is not wanted. Therfore, issue links to a copy in another project won't
get copied anymore.

Fixes #1728